requirements.txt
torch==1.12.0+cu113
torchvision==0.13.0+cu113
numpy==1.23.5
matplotlib 
tqdm
pandas
pillow
scipy
tensorboard>=1.14
pyyaml 
scikit-learn
scikit-image
h5py
wandb
torchmetrics
torcheval
seaborn
tqdm
pytorch_ood